Fetch jobs sorted by finished timestamp.
Since getStatusReport.js uses `_.find` to search through jobs, and
because `_.find` returns the first hit, make sure to tell mongo to
fetch the jobs sorted with the most recent ones first.

This way we can be sure that the status report plugin always gives a
correct status for the most recent job run for every branch per project.